Final code review thingies:
Bufferoverflow in data_buffer2wire in ldns-testpkts.c
Print unknown rcodes in ldns_axfr_next in resolver.c
Handle errors in main in ldns-keyfetcher.c
Continue in correct loop in ldns_resolver_new_frm_fp_l in resolver.c
Skip set delimeters (to del) when tokenread i.s.o. delim in ldns_bget_token and ldns_fget_token in parse.c
Test reply when verifying tsig in ldns_send_buffer in net.c
Assert that verify_next_hashed_name is only called with nsecs in the zone with ldns-verify-zone.c
Set inception and expiration on keys after they are read from the engine in main in ldns-signzone.c
Gracefully return from ldns_dnssec_zone_add_empty_nonterminals in a broken rbtree.
Check if tree->rr is null before use in ldns_dnssec_trust_tree_contains_keys in dnssec_verify.c
Check data for null before use in ldns_dnssec_create_nsec_bitmap in dnssec.c
Dead code in ldns_str2rdf_wks, ldns_resolver_query, examples/ldns_testpkts/data_buffer2wire, drill/work/packetbuffromfile & ldns_dnssec_verify_denial_nsec3_match.

+++ drill/drill_util.c
@@ -20,7 +20,7 @@ read_line(FILE *input, char *line, size_t len)
 	
 	char c;
 	for (i = 0; i < len-1; i++) {
-		c = getc(input);
+		c = (int)getc(input);
 		if (c == EOF) {
 			return -1;
 		} else if (c != '\n') {
